refactor(edit-engine): extract Net and MOS transaction domains
Move Power Rail authoring, Base Net merge, Connectivity Evidence lifecycle, and MOS bulk default/reconciliation edits behind two typed dispatchers. Keep physical-contact fixed-point normalization, deferred cleanup, and final Logical Net validation in the transaction coordinator.

Preserve exact object-ID collision checks, evidence pruning, bulk provenance repair, explicit-route ownership, changed-object IDs, and connectivity status transitions.

Validation: pnpm typecheck; pnpm test:local packages/edit-engine/src (246 tests); focused VDD, Ground, and MOS Bulk browser coverage (6 scenarios); git diff --check.

No test files changed because existing power, evidence, MOS, transaction, and browser contracts directly exercise the moved behavior.

Test-Impact: no-test-change — behavior-preserving Net/Power/MOS dispatcher extraction covered by 246 package tests and 6 browser scenarios.
